Installation of 64-bit SQL compact edition 4.0 SP1.

+++ hmailserver/source/Server/Common/SQL/ADORecordset.cpp
@@ -115,7 +115,7 @@ namespace HM
    {
       try
       {
-         return cADORecordset->GetRecordCount();
+         return (long) cADORecordset->GetRecordCount();
       }
       catch ( _com_error &err)
       {
